refactor: 예약 금액 타입 BigDecimal로 수정

+++ src/main/java/roomescape/payment/dto/PaymentConfirmResponse.java
@@ -1,7 +1,9 @@
 package roomescape.payment.dto;
 
+import java.math.BigDecimal;
+
 public record PaymentConfirmResponse(
-        int totalAmount,
+        BigDecimal totalAmount,
         String orderId,
         String paymentKey) {
 }
